--- src/App.js
@@ -1,12 +1,0 @@
-import React from "react";
-import "./App.css";
-
-function App() {
-  return (
-    <div className="App">
-      <h1>Weather Station</h1>
-    </div>
-  );
-}
-
-export default App;
