+++ src/components/LocationDetails.js
@@ -1,0 +1,14 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const LocationDetails = (props) => {
+  const { city, country } = props;
+  return <h1 className="location">{`${city}, ${country}`}</h1>;
+};
+
+LocationDetails.propTypes = {
+  city: PropTypes.string.isRequired,
+  country: PropTypes.string.isRequired,
+};
+
+export default LocationDetails;
