+++ src/index.js
@@ -1,12 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./styles/index.css";
+import App from "./components/App";
+import { location } from "./data/forecast.json";
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <App location={location} />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
-
