+++ src/tests/App.test.js
@@ -1,0 +1,9 @@
+import { render, screen } from "@testing-library/react";
+import React from "react";
+import App from "../components/App";
+
+test("renders learn react link", () => {
+  render(<App />);
+  const linkElement = screen.getByText(/Weather Station/i);
+  expect(linkElement).toBeInTheDocument();
+});
